remove "combined-ca" and fix bootstrapping
This CA combined CAs that are supposed to be used for both user
and server authentication. That is wrong, you want these two be
separate so that if you are an application that wants to choose
to trust server certificates, you don't automatically enable
client cert authentication from these signers, too (and vice versa).